Email verify

+++ src/include/da/payment-da.h
@@ -8,4 +8,4 @@
 #include "db.h"
 #include "result_bind.h"
 
-json_object * insert_into_payments(const char *user_id, const char *pub_id,char *m_id,char *c_id,float amount);
+json_object * insert_into_payments(const char *user_id, const char *pub_id,char *m_id,char *c_id,int amount);
